Add search methods from scripts to domUpdates
