+++ src/domUpdates.js
@@ -225,6 +225,27 @@ let domUpdates = {
     </ul>
     <h5>This recipe generally costs: $${cost} </h5>
     </div>`;
+  },
+
+  filterRecipes: () => {
+    const cardArea = document.querySelector('.all-cards');
+    var searchTerm = event.target.value.toLowerCase();
+    if (searchTerm !== '') {
+      cardArea.innerHTML = '';
+      domUpdates.displayFilteredRecipes(searchTerm);
+    } else {
+      domUpdates.populateCards(domUpdates.cookbook.recipes);
+    }
+  },
+
+  displayFilteredRecipes: (searchTerm) => {
+    let allRecipeResults = domUpdates.cookbook.recipes.filter((recipe)=>{
+      let ingredientNames = recipe.ingredients.map((ingredient)=>{
+        return ingredient.name
+      })
+      return recipe.name.toLowerCase().includes(searchTerm) || recipe.tags.includes(searchTerm) || ingredientNames.includes(searchTerm)
+    })
+    domUpdates.populateCards(allRecipeResults)
   }
 };
